Add links via label
